move event list to own component and memo it
